refactor: convert CheckboxGroup to RecommendationTemplateCard
